Undo moving of "concretization_version_order"
