+++ lib/spack/spack/version/version_types.py
@@ -1332,20 +1332,3 @@ def ver(obj: Union[VersionType, str, list, tuple, int, float]) -> VersionType:
         return from_string(str(obj))
     else:
         raise TypeError("ver() can't convert %s to version!" % type(obj))
-
-
-def concretization_version_order(
-    version_info: Tuple[Union[GitVersion, StandardVersion], dict],
-) -> Tuple[bool, bool, bool, bool, Union[GitVersion, StandardVersion]]:
-    """Version order key for concretization, where preferred > not preferred,
-    not deprecated > deprecated, standard version > git version, finite > any infinite component;
-    only if all are the same, do we use default version ordering.
-    """
-    version, info = version_info
-    return (
-        info.get("preferred", False),
-        not isinstance(version, GitVersion),
-        not version.isdevelop(),
-        not version.is_prerelease(),
-        version,
-    )
